0.10.5

+++ db/migrate/012_create_key_audit_logs.rb
@@ -2,8 +2,8 @@ class CreateKeyAuditLogs < ActiveRecord::Migration[5.2]
   def up
     unless table_exists?(:key_audit_logs)
       create_table :key_audit_logs do |t|
-        t.integer :key_id, null: false, index: true
-        t.integer :user_id, null: true, index: true
+        t.bigint :key_id, null: false, index: true
+        t.bigint :user_id, null: true, index: true
         t.string :action, null: false, index: true
         t.json :fields_changed, default: "[]"
         t.json :data, default: "{}"
